btrfs-progs: Properly size the leafsize field in the mdrestore_struct struct
It's 32 bits as defined in ctree.h, but the struct had it as 64 bits.

Found using MemorySanitizer.

+++ btrfs-image.c
@@ -128,7 +128,7 @@ struct mdrestore_struct {
 	struct rb_root chunk_tree;
 	struct list_head list;
 	size_t num_items;
-	u64 leafsize;
+	u32 leafsize;
 	u64 devid;
 	u8 uuid[BTRFS_UUID_SIZE];
 	u8 fsid[BTRFS_FSID_SIZE];
